fvp: pwrc: Move to drivers/ folder
