+++ plat/arm/board/fvp/aarch32/fvp_helpers.S
@@ -6,9 +6,8 @@
 
 #include <arch.h>
 #include <asm_macros.S>
+#include <drivers/arm/fvp/fvp_pwrc.h>
 #include <platform_def.h>
-
-#include "../drivers/pwrc/fvp_pwrc.h"
 
 	.globl	plat_secondary_cold_boot_setup
 	.globl	plat_get_my_entrypoint
