+++ plat/arm/board/fvp/fvp_pm.c
@@ -10,6 +10,7 @@
 #include <arch_helpers.h>
 #include <common/debug.h>
 #include <drivers/arm/gicv3.h>
+#include <drivers/arm/fvp/fvp_pwrc.h>
 #include <lib/extensions/spe.h>
 #include <lib/mmio.h>
 #include <lib/psci/psci.h>
@@ -18,7 +19,6 @@
 #include <plat/common/platform.h>
 #include <platform_def.h>
 
-#include "drivers/pwrc/fvp_pwrc.h"
 #include "fvp_private.h"
